[rename] 일정 생성 api 분리 및 시간 변환 함수 통합
